restructure and split files
