+++ tests/conftest.py
@@ -1,3 +1,5 @@
+"""Pytest runtime configuration."""
+
 from __future__ import annotations
 
 import pytest
