+++ tox.ini
@@ -22,7 +22,8 @@ commands =
 
 [flake8]
 max-line-length = 90
-extend-ignore = S101
+extend-ignore = S101, D101, D102, D103, D105, D106, D107, TC006
+per-file-ignores = flake8_trio/visitors/__init__.py: F401, E402
 
 [coverage:report]
 exclude_lines =
